requirements.txt
requests
pandas
python-dotenv
numpy
pymssql
apache-airflow
twilio
openai